# POC: combined scale + diagonal mask infinity + soft max op